// src/app/api/model/file-upload-response.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RateLimitInfo } from './rate-limit-info';
import { UploadInstructions } from './upload-instructions';


export interface FileUploadResponse { 
    error?: string;
    expiresAt?: string;
    filePath?: string;
    publicUrl?: string;
    rateLimitInfo?: RateLimitInfo;
    uploadId?: string;
    uploadInstructions?: UploadInstructions;
    uploadUrl?: string;
}