// src/app/api/model/rate-limit-info.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RateLimitInfo { 
    remainingDaily?: number;
    remainingHourly?: number;
    storageLimitMB?: number;
    storageUsedMB?: number;
}